Extract WeakRefCollection to its own file
